Changed PK variable types from Int to Long.

+++ AgendaServer/src/main/kotlin/service/GroupService.kt
@@ -10,7 +10,7 @@ class GroupService {
         TB_Group.selectAll().map { toGroup(it) }
     }
 
-    suspend fun getGroup(id: Int): Group? = dbQuery {
+    suspend fun getGroup(id: Long): Group? = dbQuery {
         TB_Group.select {
             (TB_Group.id_group eq id)
         }.mapNotNull { toGroup(it) }
@@ -33,7 +33,7 @@ class GroupService {
     }
 
     suspend fun addGroup(newGroup: NewGroup): Group {
-        var key: Int? = 0
+        var key: Long? = 0
         dbQuery {
             key = model.TB_Group.insert({
                 it[group_title] = newGroup.group_title
@@ -43,7 +43,7 @@ class GroupService {
         return getGroup(key!!)!!
     }
 
-    suspend fun deleteGroup(id: Int): Boolean = dbQuery {
+    suspend fun deleteGroup(id: Long): Boolean = dbQuery {
         TB_Group.deleteWhere { TB_Group.id_group eq id } > 0
     }
